moved stuff around part 2
