+++ src/runtime_src/xdp/profile/plugin/aie_trace/util/aie_trace_config.cpp
@@ -37,52 +37,6 @@ namespace xdp::aie::trace {
   using severity_level = xrt_core::message::severity_level;
 
   /****************************************************************************
-   * Print out resource usage statistics for a given tile
-   ***************************************************************************/
-  void printTileStats(xaiefal::XAieDev* aieDevice, const tile_type& tile)
-  {
-    if (xrt_core::config::get_verbosity() < static_cast<uint32_t>(severity_level::info))
-      return;
-
-    auto col = tile.col;
-    auto row = tile.row;
-    auto loc = XAie_TileLoc(col, row);
-    std::stringstream msg;
-
-    const std::string groups[3] = {
-      XAIEDEV_DEFAULT_GROUP_GENERIC,
-      XAIEDEV_DEFAULT_GROUP_STATIC, 
-      XAIEDEV_DEFAULT_GROUP_AVAIL
-    };
-
-    msg << "Resource usage stats for Tile : (" << col << "," << row << ") Module : Core" << std::endl;
-    for (auto& g : groups) {
-      auto stats = aieDevice->getRscStat(g);
-      auto pc = stats.getNumRsc(loc, XAIE_CORE_MOD, XAIE_PERFCNT_RSC);
-      auto ts = stats.getNumRsc(loc, XAIE_CORE_MOD, xaiefal::XAIE_TRACE_EVENTS_RSC);
-      auto bc = stats.getNumRsc(loc, XAIE_CORE_MOD, XAIE_BCAST_CHANNEL_RSC);
-      msg << "Resource Group : " << std::left << std::setw(10) << g << " "
-          << "Performance Counters : " << pc << " "
-          << "Trace Slots : " << ts << " "
-          << "Broadcast Channels : " << bc << " " 
-          << std::endl;
-    }
-    msg << "Resource usage stats for Tile : (" << col << "," << row << ") Module : Memory" << std::endl;
-    for (auto& g : groups) {
-      auto stats = aieDevice->getRscStat(g);
-      auto pc = stats.getNumRsc(loc, XAIE_MEM_MOD, XAIE_PERFCNT_RSC);
-      auto ts = stats.getNumRsc(loc, XAIE_MEM_MOD, xaiefal::XAIE_TRACE_EVENTS_RSC);
-      auto bc = stats.getNumRsc(loc, XAIE_MEM_MOD, XAIE_BCAST_CHANNEL_RSC);
-      msg << "Resource Group : " << std::left << std::setw(10) << g << " "
-          << "Performance Counters : " << pc << " "
-          << "Trace Slots : " << ts << " "
-          << "Broadcast Channels : " << bc << " " 
-          << std::endl;
-    }
-    xrt_core::message::send(severity_level::info, "XRT", msg.str());
-  }
-  
-  /****************************************************************************
    * Configure stream switch ports for monitoring purposes
    * NOTE: Used to monitor streams: trace, interfaces, and memory tiles
    ***************************************************************************/
